git-svn-id: https://svn.code.sf.net/p/miktex/code/miktex/trunk@2233 3fc76331-e2db-4b81-af9c-3600f57961af

+++ Programs/TeXAndFriends/xetex/XeTeXFontInst.h
@@ -58,7 +58,7 @@ authorization from the copyright holders.
 
 #if ! defined(MIKTEX)
 extern "C" {
-	void *xmalloc(unsigned);	// from kpathsea
+	void *xmalloc(size_t);	// from kpathsea
 };
 #endif
